pyproject.toml
[build-system]
requires = ["hatchling>=1.18"]
build-backend = "hatchling.build"

[project]
name = "last-words"
version = "0.2.0"
description = "A conversational puzzle game where the player talks a sentient AI bomb out of detonating."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "Proprietary" }
authors = [
    { name = "Last Words Team" },
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

# Core runtime dependencies. Intentionally minimal; provider-specific SDKs
# are optional extras so the core package stays model-agnostic.
dependencies = [
    "pyyaml>=6.0",
]

[project.optional-dependencies]
# Install with:  pip install -e ".[anthropic]"
anthropic = ["anthropic>=0.80.0"]
# openai     = ["openai>=1.0.0"]        # add when OpenAIProvider is implemented
# google     = ["google-generativeai"]  # add when GoogleProvider is implemented
# local      = ["llama-cpp-python"]     # add when LocalProvider is implemented
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.1.0",
    "mypy>=1.5",
    "pre-commit>=3.0",
]

[project.scripts]
last-words = "last_words.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/last_words"]

[tool.hatch.build.targets.sdist]
include = [
    "src/last_words",
    "prompts",
    "config",
    "README.md",
    "CHANGELOG.md",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
markers = [
    "integration: tests that require a live LLM API key",
    "unit: pure unit tests with no external dependencies",
]
addopts = [
    "-ra",
    "--strict-markers",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "N", "SIM"]
ignore = ["E501"]  # line length handled by formatter

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true
# Third-party libraries without stubs:
[[tool.mypy.overrides]]
module = ["anthropic.*", "yaml.*"]
ignore_missing_imports = true
